+++ BetterChests/Features/CraftFromChest.cs
@@ -1,13 +1,10 @@
 namespace StardewMods.BetterChests.Features;
 
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using StardewModdingAPI.Events;
-using StardewModdingAPI.Utilities;
 using StardewMods.BetterChests.Helpers;
 using StardewMods.BetterChests.Models;
-using StardewMods.BetterChests.StorageHandlers;
 using StardewMods.Common.Enums;
 using StardewMods.Common.Integrations.BetterChests;
 using StardewValley.Locations;
@@ -18,16 +15,10 @@ using StardewValley.Menus;
 /// </summary>
 internal class CraftFromChest : IFeature
 {
-    private const int MaxTimeOut = 60;
-
     private static CraftFromChest? Instance;
 
     private readonly ModConfig _config;
-    private readonly PerScreen<int> _currentTab = new();
-    private readonly PerScreen<List<IStorageObject>> _eligible = new(() => new());
     private readonly IModHelper _helper;
-    private readonly PerScreen<List<IStorageObject>> _locked = new(() => new());
-    private readonly PerScreen<int> _timeOut = new();
 
     private bool _isActivated;
 
@@ -47,25 +38,9 @@ internal class CraftFromChest : IFeature
            && storage.Source is not null
            && storage.CraftFromChest.WithinRangeOfPlayer(
                   storage.CraftFromChestDistance,
-                  storage.Source,
+                  storage.Location,
                   storage.Position)
         select storage;
-
-    private List<IStorageObject> CurrentEligible => this._eligible.Value;
-
-    private int CurrentTab
-    {
-        get => this._currentTab.Value;
-        set => this._currentTab.Value = value;
-    }
-
-    private List<IStorageObject> LockedEligible => this._locked.Value;
-
-    private int TimeOut
-    {
-        get => this._timeOut.Value;
-        set => this._timeOut.Value = value;
-    }
 
     /// <summary>
     ///     Initializes <see cref="CraftFromChest" />.
@@ -87,8 +62,7 @@ internal class CraftFromChest : IFeature
         }
 
         this._isActivated = true;
-        this._helper.Events.GameLoop.UpdateTicked += this.OnUpdateTicked;
-        this._helper.Events.GameLoop.UpdateTicking += this.OnUpdateTicking;
+        this._helper.Events.Display.MenuChanged += CraftFromChest.OnMenuChanged;
         this._helper.Events.Input.ButtonsChanged += this.OnButtonsChanged;
 
         if (Integrations.ToolbarIcons.IsLoaded)
@@ -98,7 +72,7 @@ internal class CraftFromChest : IFeature
                 "furyx639.BetterChests/Icons",
                 new(32, 0, 16, 16),
                 I18n.Button_CraftFromChest_Name());
-            Integrations.ToolbarIcons.API.ToolbarIconPressed += this.OnToolbarIconPressed;
+            Integrations.ToolbarIcons.API.ToolbarIconPressed += CraftFromChest.OnToolbarIconPressed;
         }
 
         if (Integrations.BetterCrafting.IsLoaded)
@@ -116,14 +90,13 @@ internal class CraftFromChest : IFeature
         }
 
         this._isActivated = false;
-        this._helper.Events.GameLoop.UpdateTicked -= this.OnUpdateTicked;
-        this._helper.Events.GameLoop.UpdateTicking -= this.OnUpdateTicking;
+        this._helper.Events.Display.MenuChanged -= CraftFromChest.OnMenuChanged;
         this._helper.Events.Input.ButtonsChanged -= this.OnButtonsChanged;
 
         if (Integrations.ToolbarIcons.IsLoaded)
         {
             Integrations.ToolbarIcons.API.RemoveToolbarIcon("BetterChests.CraftFromChest");
-            Integrations.ToolbarIcons.API.ToolbarIconPressed -= this.OnToolbarIconPressed;
+            Integrations.ToolbarIcons.API.ToolbarIconPressed -= CraftFromChest.OnToolbarIconPressed;
         }
 
         if (Integrations.BetterCrafting.IsLoaded)
@@ -132,14 +105,20 @@ internal class CraftFromChest : IFeature
         }
     }
 
-    private void ExitFunction()
+    private static void OnMenuChanged(object? sender, MenuChangedEventArgs e)
     {
-        foreach (var storage in this.LockedEligible)
+        if (e.NewMenu is GameMenu)
         {
-            storage.Mutex?.ReleaseLock();
+            BetterCrafting.AddMaterials(CraftFromChest.Eligible);
         }
+    }
 
-        this.LockedEligible.Clear();
+    private static void OnToolbarIconPressed(object? sender, string id)
+    {
+        if (id == "BetterChests.CraftFromChest")
+        {
+            BetterCrafting.ShowCraftingPage(CraftFromChest.Eligible);
+        }
     }
 
     private void OnButtonsChanged(object? sender, ButtonsChangedEventArgs e)
@@ -150,106 +129,6 @@ internal class CraftFromChest : IFeature
         }
 
         this._helper.Input.SuppressActiveKeybinds(this._config.ControlScheme.OpenCrafting);
-        this.OpenCrafting();
-    }
-
-    private void OnToolbarIconPressed(object? sender, string id)
-    {
-        if (id == "BetterChests.CraftFromChest")
-        {
-            this.OpenCrafting();
-        }
-    }
-
-    private void OnUpdateTicked(object? sender, UpdateTickedEventArgs e)
-    {
-        if (this.TimeOut == 0 || --this.TimeOut != 0)
-        {
-            return;
-        }
-
-        foreach (var storage in this.CurrentEligible)
-        {
-            storage.Mutex?.ReleaseLock();
-        }
-
-        this.CurrentEligible.Clear();
-        this.TimeOut = 0;
-        var width = 800 + IClickableMenu.borderWidth * 2;
-        var height = 600 + IClickableMenu.borderWidth * 2;
-        var (x, y) = Utility.getTopLeftPositionForCenteringOnScreen(width, height).ToPoint();
-        Game1.activeClickableMenu = new CraftingPage(
-            x,
-            y,
-            width,
-            height,
-            false,
-            true,
-            this.LockedEligible.OfType<ChestStorage>().Select(storage => storage.Chest).ToList())
-        {
-            exitFunction = this.ExitFunction,
-        };
-    }
-
-    private void OnUpdateTicking(object? sender, UpdateTickingEventArgs e)
-    {
-        foreach (var storage in this.CurrentEligible)
-        {
-            storage.Mutex?.Update(storage.Source as GameLocation ?? Game1.currentLocation);
-        }
-
-        if (Game1.activeClickableMenu is not GameMenu { currentTab: var currentTab } gameMenu
-         || currentTab == this.CurrentTab)
-        {
-            return;
-        }
-
-        this.CurrentTab = currentTab;
-        if (gameMenu.pages[currentTab] is not CraftingPage craftingPage)
-        {
-            return;
-        }
-
-        craftingPage._materialContainers ??= new();
-        craftingPage._materialContainers.AddRange(
-            CraftFromChest.Eligible.OfType<ChestStorage>().Select(storage => storage.Chest));
-        craftingPage._materialContainers = craftingPage._materialContainers.Distinct().ToList();
-    }
-
-    private void OpenCrafting()
-    {
-        this.CurrentEligible.Clear();
-        this.CurrentEligible.AddRange(CraftFromChest.Eligible);
-        if (!this.CurrentEligible.Any())
-        {
-            Game1.showRedMessage(I18n.Alert_CraftFromChest_NoEligible());
-            return;
-        }
-
-        if (Integrations.BetterCrafting.IsLoaded)
-        {
-            Integrations.BetterCrafting.API.OpenCraftingMenu(
-                false,
-                false,
-                null,
-                null,
-                null,
-                false,
-                this.CurrentEligible.Select(
-                        storage => new Tuple<object, GameLocation>(
-                            new StorageWrapper(storage),
-                            storage.Source as GameLocation ?? Game1.currentLocation))
-                    .ToList());
-            return;
-        }
-
-        this.TimeOut = CraftFromChest.MaxTimeOut;
-        for (var index = this.CurrentEligible.Count - 1; index >= 0; index--)
-        {
-            var storage = this.CurrentEligible[index];
-            storage.Mutex?.RequestLock(
-                () => this.LockedEligible.Add(storage),
-                () => this.CurrentEligible.Remove(storage));
-        }
+        BetterCrafting.ShowCraftingPage(CraftFromChest.Eligible);
     }
 }